refactor(meeting,history): extract export formatters and history filter state (#699, #703)
#699: Move MeetingExportFormat enum + 5 pure formatting functions
(meeting_session_text, meeting_session_csv, meeting_session_json,
rendered_speaker_label, format_relative_timestamp) from
ipc/commands/meeting.rs to meeting/export.rs — the IPC layer was
the wrong home for domain transformations with zero IPC dependencies.
ipc/commands/export.rs was also importing from the old location;
updated to crate::meeting::export. Tests moved with the code.
meeting/mod.rs re-exports MeetingExportFormat so IPC callers still
get it through the domain crate root.

#703: Move HistoryFilter type, FeedRow type, historyFilter state,
effectiveFilter derived, and mergedFeed two-pointer derived into
history.svelte.ts. HistoryPanel.svelte is now a thin presenter:
reads history.filter / history.effectiveFilter / history.mergedFeed,
writes filter via history.filter = next. exportBundle no longer needs
activeFilter passed in from the panel — it reads effectiveFilter from
module scope.
